refactor!: Move freezing code to workspace.freezer
